Big update, read the description.
Okay, so I forgot 2 key VTF's that were key for good phong and good looking textures.

These have been added along with an experimental version of the Lee Enfield, but scoped.

Textures and phong should look A LOT better (this doesn't count for the shotgun).

The Lee Enfield is WIP.